WT-12036: Add previously added style classes of `WAbstractToggleButton`
This class is a bit special, in that is recreates its DomElements
different, since it can have inputs, labels and additional text.

These items are not completely "persisted", and thus when recreating
them, some properties need to be copied over. Any custom style classes
as well.

Note that this is only the case for Bootstrap 5. Any other Bootstrap
theme or CSS theme does not suffer from this, since they do not add
other style classes to the element.