WT-12006: More robustly place the JS member functions
When placing a JS member on a widget, its name and value are remembered.
This is so that if the value is cleared or replaced, we do so for the
same member function.

There, however, exists a special case for member function where if a
single character is assigned to it, it is replaced by an empty value.
This serves to function such that setting the function to `0` for
example, will result in it being empty.

When we appended the semicolon to it, this caused that last piece of
logic to be ignored, and would generate incorrect JS, which called a
function `0;`.