WT-12143: Avoid adding semicolons to a self-invoking function
While adding semicolons, to ensure that every statement is correctly
closed, we were being too eager. There was the possibility that a
declared member JS function would be used as a self-invoking function.
This function was then terminated by a semicolon, leading to a
`(function(...){};)(...);` call, instead of `(function(...){})(...)`.

The semicolon is being dropped in this special case.