WT-11705: Add a process to manage TOTP "login"
This process will be shown if a second factor authentication is
required. If so, the user will first be shown a QR code and the secret
key, which they can save in any authenticator app or extension. Then
they will be required to enter the TOTP code generated from it.

Subsequent logins will no longer show the QR code or secret key, but
only request the TOTP code.