WT-11705: Show the TotpProcess after using an environment token
This token can be either an email token or a cookie. In case of an
email, the user will have registered before, and be sent a link. This
link takes them to the page where they can verify their email. Once
verified, they will be shown the TotpProcess.

For cookies, it works similar. The user's password has been remembered,
but they will still be prompted with the TOTP code request.