WT-12158: Correctly set the path for the multi-session cookies
This corrects an oversight from WT-7814, where the Http::Cookie class
was introduced. Before if the path of the cookie was set to an empty
string, the path was set to the deployment path. This lead to a wrong
substitution where the path of the new cookie class was never set, and
the path of the cookie would always be set to empty in the browser.

This results in colliding cookies between different deployments, and
would confuse the matching of sessions on a cookie basis (when working
with combined tracking).