WT-12174: Create Selenium script to automatically test widget gallery
This script allows you to run Selenium with either Firefox or Chrome, in
a headless manner. It will go over all menu items of the widget gallery
and open each in turn. It will check if no JS error occurred and no
alerts happened.

For Firefox, specialized functionality had to be created to retrieve the
logs of the browser. First an extra configuration is set that puts all
devtools output on stdout. This is logged to a file, and the file is
queried for the output, to see whether an issue occurred. This is
because the W3C spec does not have a `get_log` yet, which Chrome does
provide.

If the script completes successfully, the output will be 0, otherwise 1.