WT-12148: Allow WApplication to set attributes for the body and html
There are now setters and getters for attributes that can be added to
the main <html> and <body> tags of the application.

Before, only style classes were possible, but now any attribute can be
explicitly set. This essentially means that "class", "lang", and "dir"
can be controlled as well, but this should be preferably be avoided.

An example use-case here can be to for example enable or disable
translations or browser spellchecking.