WT-12148: Set html and body attributes on progressive bootstrap pages
Pages that are updated with JavaScript need to set the html and body
attributes by calling JavaScript. For the body it sets the attributes on
`document.body`, while for the html it does so for
`document.documentElement`.

This will then set the attributes both when the page is initially loaded
(together with the Wt JavaScript framework), and when any update
happens, if any of the values have changed.