WT-11685: Do not use `setHtml` in JavaScript for updating widgets
When this is used, specifically for widgets that can have children, it
may remove children that were added before. Specifically this happens
when the children were not added in C++ code, but via JavaScript code.

An example is when a resize sensor is added to a `WContainerWidget`,
which can be achieved by setting `setLayoutSizeAware(true)`. The resize
sensor is added to the container purely in JavaScript, which is not
reflected in C++ code. The amount of children the WContainerWidget has
there remains 0. If a child is then added to it, it is done so via
`setHtml`, which overrides the resize sensor.

We can avoid this by not allowing this to happen if the widget is being
updated.