WT-12036: Do not rely on `std::find` (for JWt)
The translation tool we use for generating JWt's code cannot correctly
translate an `std::find`. That's because it would require the actual
vector to be passed as an argument, not its iterators.

We cannot choose to manipulate the content of the function conditionally
(e.g. if the vector iterator is found inside an `std::find` call, we do
a different substitution). As such we revert to a more "primitive"
approach, going over the elements manually.

This can be avoided were we to use a custom function that generates
ranges, or use something like `std::ranges`.