WT-11772: Resolve destructor warnings
There were missing virtual destructors. Whenever a class has virtual
functions, and it used as a base class, it is required to have a virtual
destructor.

Additionally, `Wdeprecated-copy-with-user-provided-copy` errors were
generated, due to a bad application of the rule of zero/three/five.
There was a user-provided copy assignment operator. As such, we need
to also define a destructor.