WT-11705: Add a base32 implementation
This implementation features a function to both decode and encode base32
strings. The encoding algorithm is based on Ryan Petrie's implementation
of the base64 encoder. The decoder is not. That is a custom
implementation. Otherwise, if the number of encoded groups is larger, the
number of specific cases we need to cover for the eventual result shifting
would be much too cumbersome. This would result in a big number of cases
and many lines of code to handle them. So we moved to a system that simply
decodes the character whenever enough bits are present (8).