WT-12145: Add a missing `flexDirection` JavaScript name
This JavaScript name was missing, and thus resulted in a badly indexed
retrieval, mismatching the element and the name. Meaning for all
elements following this index, the index was off by one, making the
value it attached and the name of the style wrong. This only effected a
couple value, namely "flex-flow", "flex-direction", "align-self", and
"justify-content", If they are set using the internal `Wt::Property`
enum class.