WT-12145: Replace the `DomElement` CSS names and JavaScript names arrays
These were simply arrays, that did nothing when a wrong index would be
retrieved. E.g. if the arrays were not of the right size (and were for
example missing a value), a wrong name would be retrieved, instead of it
warning that the name was missing.

This lead to this bug, where a `flexDirection` was missing from the
JavaScript CSS properties array.