WT-12145: Extract `unorderedmap` access and insert functionality
This ensures that there is a unified interface, that can also be used by
JWt. Additionally, the construction of the map is extracted to a
function, to ensure that the substitution happens correctly after
parsing the AST. Otherwise, if it isn't an immediate static member, it
will be converted to such in Java, but too late to be correctly caught
by the mapper filler, leading to wrong methods existing.