WT-11705: Enabled "remember-me" functionality for MFA verification
It would be annoying to users to ask them for a TOTP code each time they
log in (if MFA is enabled, and either required or a secret is present in
the user's record "auth_info"). We can allow the auth tokens that are
currently only used for "regular" login (username/password) to be
extended with a type that can be used to remember MFA verification steps
as well. The name of the cookie, and its validity are "wtauth-mfa" and
one week by default respectively. They can both be changed from the
`AuthService`.

This does change the "auth_info" record, and "auth_token" records. And
should be migrated on existing databases. This will be explicitly and
clearly mentioned in the release notes.