WT-12352: Resolve a floatingpoint exception in the Boost spirit grammar
We use a grammar parser to evaluate the plural parameters of
`WMessageResource`s. A simple grammar was written to facilitate this.

These expressions are managed by the developer, and not really a viable
attack vector. However, given a parameter string that contains an unsafe
division or modulo (dividing by 0), there is the possibility of a crash.

This was noticed by the recently added OSS Fuzz (#63734).