WT-9284: Set the size of `WTableView` correctly
The size of a `WTableView` was never correctly set. The whole chain of
resize calculations did give a somewhat correct approximation, but due
to timing issues, on when exactly the `wtResize` was executed, and when
the whole layout adjustment took place, a different result was achieved.

Either the table fit the wrapping layout completely, resulting in a
table that grew as large as the surrounding layout. Or it shrunk to fit
its content. The former happened more often, but the latter did occur
regularly.

To void this, we check when adjusting the layout (which would prefer the
shrunk table), if the table is managed by a layout manager. If it is, we
leave the size calculation alone.