WT-9284: Size the contents of a `WTableView` correctly
The content div is where the scrollbars are placed, and are also used to
decide if scrollbars are indeed to appear. Before, this was not set
accurately, resulting in a table where the scrollbar was floating in the
middle, if its content was set to be of a certain size.

For a table that had a very simply layout wrapper, the table contents
could grow outside of its bounds.