WT-10382: Add a customized WFileDropWidget example
This adds another example for WFileDropWidget, based on the `filedrop`
example, that shows how to customize the upload mechanism.

The given example is based on the existing `filedrop` example. It only
simplifies the existing upload mechanism and adds some logging, but it
should give enough information to use it to build more complex data
transfer mechanisms. (At the client side it sends the data to a third
party to at least do something interesting.)