WT-12374: Introduce a new derived exception for invalid operations
An invalid operation can be defined as many things, like calling
functions in  a bad order, or performing the call on bad arguments,
leading to a failing call (like dividing by 0) or unidentified behaviour.

If you defensively check the argument a `WInvalidArgumentException`
would probably be much more appropriate. But this is currently not yet
added. This will likely be created soon though.

This again allows developers to have better control over the exceptions
they can handle in their code.