WT-12374: Temporarily impose a limit to avoid stack overflows
When boost is parsing the evaluation, it will apply recursive rules, and
push items to the stack. At a certain point, this may become too full,
causing it to overflow. The better solution here would be to ensure that
we limit the actual depth of the expressions that can be evaluated.

However, I did not find a way to do this reliable in Boost spirit classic
without side-effects (e.g. eps_p). It seems that Boost spirit Qi does
allow this. The conversion from one to the other is giving some issues,
and since this is mainly to avoid Fuzzer issues, I decided that this
wasn't worth it for now.

The current limit is set somewhat arbitrarily. The exception seems to
occur after roughly 1500 character of the input. Other inputs may
exhibit the earlier or later, so for safety 1000 was chosen.