WT-11969: Add better building documentation for dependencies
There were different indication of minimal and core dependencies all
over the place. Now they have been added in
`doc/MinimumDependencyVersions.md`, as a single source of truth.

The Doxygen documentation will include segments of this file. All
information can be managed from that single file.

The file gives all core dependencies, and all optional dependencies that
Wt can be build with. For each of these, the most common compilation
flags have also been explained. For most this is a simple ON/OFF switch,
and an optional PREFIX in case it isn't a system-wide installation.

Note that the Windows building documentation is still found on Redmine,
and severely outdated, and is to be converted at some point in the
future.