WT-12461: Allow headers to be added/inserted multiple times
Currently only the function `addHeader` exists, which doesn't allow the
header to be replaced, nor cleared. Once added, the header will keep
existing. This can result in multiple headers with the same name
existing, or headers with empty values being present.

We now introduce `insertHeader`, which allows the header to be replaced
if it already exists, or cleared if an empty value is passed to it.

This is necessary to ensure that the protocols and extensions
`WWebSocketResource` will return in its handshake, can both be empty.