WT-12461: Allow sockets to be transferred from session connections
Connections that have been made as part of a session, to dynamic
resources can now be transferred. This will allow `WWebSocketResource`s
that are put in place in a private manner (bound to the session id), to
transfer their socket, and allow the socket to function as the
communication channel for the `WWebSocketResource` directly as a
`WWebSocketConnection`.