WT-12461: Allow a `WApplication` to manage `WWebSocketResource`s
These are special resources, similar to the normal resources. However,
they have an additional layer here. A `WWebSocketResource` is not a
`WResource`, that would make the class way too big, and already put it
in an awkward spot, having to adhere to how `WResource`s work.

As such, a `WWebSocketResource` has a private `WResource`that actually
handles the initial handshake, which is registered to the `WApplication`
so that the session can find it. Once the handshake has occurred, the
socket is then transferred to the resource itself, and will be managed
outside of the normal flow.