WT-12461: Set up the ping-pong system for `WWebSocketConnection`
A connection is kept alive by means of the ping-pong system.
Any of the connected parties can send a Ping frame at any time, to which
the other has to respond with a Pong frame.

The server side (`WWebSocketConnection)` is allowed to configure two
intervals. The first is the actual ping interval, denoting every how
many seconds a Ping frame is sent from the server. If set to 0 this will
not take effect.

The second in the timeout, denoting after how much time of no response
being received, the connection will be considered dead. Again, a value
of 0 will denote this not being used.