WT-12461: Enable `WWebSocketResource` to take the application lock
Taking the lock means that the functionality that is executed inside the
lock is performed on the main application thread safely, outside of the
main event loop.

Allowing the resource to take this enabled to developer not to consider
side-effects/safety of adding information to the UI. It's essentially a
convenience to ensure that UI updates (or more general: updates to any
and all widgets and their application) can be performed without much
thought.