WT-12461: Allow `WebRequest` to transfer its socket
When a WebSocket request is being set up, its socket needs to be
transferred from the http library to the WWebSocketConnection class.

When a connection to the server is opened, and it matches the right
filters, identifying it as a non-framework WebSocket request, the
connection is marked as one where socket ownership will be transferred.

This happens when the connection would normally be closed. A callback is
registered to the connection when this happens, which moves the socket
to the WebSocket implementation.