WT-12538: Put multiple recipients of the same type on the same line
When multiple recipients of the same type ("To" or "Cc") are added to a
message, they were each put on a separate line, possibly resulting in
multiple lines with the "To: " or "Cc: " headers. This is not in
compliance with the RFC5322 specification. It states that only a single
line with a comma-separated list of addresses should be used.

It seems that very few services complained about this, but one that does
follow the spec, could have rejected messages generated by Wt.