WT-12606: add directory upload to WFileDropWidget
This is supported by the introduction of the Directory class, which
inherits from File. So if a 'File' is dropped, this can now also be a
Directory. The Directory contains a vector of Files (which could again be
Directories).

The files that are actually uploaded are still stored in a single vector,
so directories are 'flattened' into this vector.

If a file is uploaded as part of a directory, it will also have the path()
property which is the path relative to the directory that was dropped.