WT-11705: Add MFA configuration to `AuthService`
MFA set-up and configuration is now managed by `AuthService`, where the
MFA functionality can be enabled/disabled, can be set as required or
not, and the length of the code that is to be validated can be set.

When this is enabled, it is not automatically enforced for all users,
rather then it is up to the developer to ensure that it can be set for
each user separately. If the provider/identity is configured, it will be
automatically picked up by Wt, and be made part of the login
functionality. On the other hand, if it is marked as required, it will
be enforced for all users, and will be shown to users if they do not
have it configured.