WT-12461: Pass the WebSocket requests correctly through the main process
If the server allows dedicated sessions, there is essentially a main
process that will direct incoming connections to the right process, via
the `ConnectionManager`. The WebSocket connection however is also parsed
and interpreted, resulting in an incorrect WebSocket connection
creation. We avoid doing this, in case we are on the "proxy". The actual
child process that will inherit the connection is allowed to perform the
upgrade and send that to the connecting client.