WT-12461: Set the right origin
In the example, the environment had already upgraded to WebSocket as its
URL scheme, due to the main framework WebSocket connection being used.
So it has to take that one into account when selecting the right schema
to create the new connection.

Additionally, when verifying whether the origin is trusted in the
WebSession, we also take the possibility of WebSockets into account.
Meaning that a secure channel can only happen from another secure
channel (be that "https" or "wss").