WT-11954: Resolve the Pango memory leak in the font map
The font map is currently a global variable, since thread cleanup is a
concern. This is to be investigated later as well. Additionally since
the font map operations are not clearly stated to be thread-safe, it is
safer to only construct the map once, and then keep it in memory.

We now introduce a custom deleter, that deallocates the object upon
going out of scope. This now unreferences the GLib object correctly,
clearing the memory it uses. This reduces the total leaked amount
related to Pango significantly (though not yet completely).