WT-11954: Ensure proper Pango memory management though smart pointer
Pango calls always return raw pointers, which we used to track as simply
raw pointers as well. However, the clean-up of these pointers did not
occur always. Pointers could be created, and not cleaned up if they
would be removed from the cache before the actual destructor ran, which
happened often.

By switching to smart pointers, we ensure that the cache always remains
in a state where it managed the lifetime of fonts and their
descriptions. Upon deleting, the custom deleter is called, which will
free or unreference the Pango or GLib object.

This ensures that a lot less leaks take place. From some simple tests
the leaked size was reduced from being in the 100's of KB's to just a
few KB's.