WT-11954: Allow the font cache to reach 10 fonts
Fonts can be reused, implemented with an LRU system, so that recently
used fonts are matched sooner, by storing them at the beginning of the
list.

A fairly arbitrary limit of 10 fonts is now supported, instead of
previously popping fonts way too soon, essentially invalidating the
usage of a cache.