WT-12368: fix small input box in widgetgallery
In the forms/validation section of the widgetgallery, the WLineEdit
would be to smaal when using the theme bootstrap 3, making it impossible
for the user to see what was written in it.

We changed replaced the div class input-group in the templates by the
div class row which are wrappers specifically designed for the columns
we use.