WT-12606: directory selection through a browser dialog
Since it is only possible to open either a file picker or a directory
picker, the choice was made to keep clicking the widget a file picker, even
if directories are accepted.

To still allow users to select directories with a dialog, it is now possible
to programmatically open the file/directory selection dialog.