WT-12283: Do not count an empty JS call as a manipulation
When an update happens to an element, we collect JS to perform the
update on the browser side. We hold a counter to indicate how many
changes need to happen per element. For some very simple changes, we
short-cut to avoid processing each element in depth.

An early return when calling JS if the JS was empty, still wrongly
increased this counter.