WT-12934: Make resize consider all layout's children
When increasing the size of a widget in a layout, the maximum increase
of size was dictated only by the size and minimum size of the item
widget after the resizable one. However, the area loss would be shared
by all the widgets.

Take into account all the widgets in the layout when resizing. This
allows to resize a widget to it's maximum size in a single resize.