WT-12934: make maxSize impact resize in layout
If a widget in a layout had a maximum size, this maximum size would not
be taken into account when resizing the widget. The size of the widget
would not be bigger than it's max size, but the other widget would still
be moved away from it.

Now the handle is blocked when dragging it further would result in the
widget taking more place than it's max size.