WT-12606: improve boolean attribute value
this was working, but according to MDN, the value of a boolean attribute
can only be empty or equal to the name of the attribute.
The main reason I'm changing it is because I noticed that if the value is
set to 'false', it is still true ... The way to disable a boolean attribute
is to remove it.