WT-12951: Fix incorrect layout in WPanel
Fix an issue where a layout in a WContainer in a WPanel would consider
the min-width of the WContainer as the width of the container during an
animation if the WContainer had no given size.

This is solved by explicitly setting the width of the WContainer (or the
widget in it's place) during the animation.