WT-12995 add WLeafletMap::Popup
Add the possibility to add popups to a WLeafletMap.

This also adds the a new LeafletMapItem class from which both Popup and
Marker inherit. This class is where the communication with WLeafletMap
is handled to avoid too much code duplication. Functionality that was
previously handled by the Marker class itself, has been moved here.