WT-12995: fix setContent with hidden popup crash
When setting the content of a hidden WLeafleMap::popup, the application
would crash if the content had to call js after initialisation (like
when using a js layout). The crash happens because the content in the
popup cannot be found by the application when the popup is hidden.

To solve this problem, if the popup is closed, we open the popup before calling the
after-initialization js and close the popup just after.