WT-13034: Add a basic pipeline for Wt CI/CD
This pipeline is to be integrated with Gitlab. This allows us to build
remotely, and automatically run some tests.

A code formatting check will be appended to it later on (WT-12206).

Additionally, as a later step, if this pipeline is successful, it can
start another job, building the `wt-port` equivalent of the branch
(WT-13038).

This pipeline will be running on `builder1` (labeled `docker`), whereas
all others will run on `builder2` (labeled  `wt`).

The pipeline ought to check:
- JS formatting & linting
- Code formatting
- Build success
- Test results