WT-13034: Update Gitlab statuses in pipeline
The pipeline now updates Gitlab when it starts to run a stage, or when
it fails/succeeds.

Note that `post` can be used after a specific stage (like a build), or
after a set of stages. When running as the last part of a stage, it
should be specified at the end of the stage's steps.