WT-12866: add delay-load-at-boot option
In some very specific circumstances, the loading of the application
would take several seconds at boot time. This seems to be due to the
delay added at loadtime.

As skipping that delay might impact some of the users, we added an
option to choose whether to use the delay or not, keeping the delay by
default.