WT-13038: Add a specific agent to build CNOR
A separate dockerfile is used for each of these agents. This is a little
cleaner than using the same overblown dockerfile.

Additionally this will make moving to different builders easier if we
were to ever do that. The functionality of each step is more cleanly
separated.

This does necessitate the creation of two main steps `wt` and `wt-port`.
These each have their own `post`, which cleans up their workspace on the
builder. It is necessary to do so, since a "global" post will be
executed in `agent none`, causing an error.