WT-11705: Add throttling to MFA
This relies on a more custom approach, and will ask for developers to
consider their implementation as well. Since now standardised UI is
provided for MFA, as the process is allowed to be nearly any
implementation, throttling is to be integrated manually.

This has been elucidated in the tutorial, and documentation. The default
TOTP approach has throttling implemented, but not activated by default.