WT-993: allow fast usage of notify for postgresql
With postgresql, it is not possible to send querry while another one is
already active. This would make the application crash if multiple notify
or listen were done at the same time.

This solve the problem by storing the notify and listen if the
connection is already in use. This also implies that those notification
must be sent before destroying the postgresql connection.