WT-13034/WT-13038: Ensure that cancelled steps are caught in Gitlab
When a build was aborted, which can be done manually, or happen when a
new push to the same branch supersedes the original build, the build is
aborted. But this state was never caught in the pipeline to update
Gitlab. These pipeline states are still "Running" on Gitlab.

Now, each step that can encounter an abort, propagates this cancelled
pipeline correctly to Gitlab.