WT-13034/WT-13038: Ensure that superseded pipelines are canceled
Superseded pipelines do not ever get in the `aborted` post stage. They
are a special case. The "fix" here is to look at the final result.

In case of a build failure, this will be 'FAILURE', and in case of a
supersession 'NOT_BUILT'. This latter event is passed to Gitlab as a
canceled pipeline.