WT-13106: Ensure the right disabled theme class is used
This follows the work done in WT-10512. Here the disabled state of
certain composite widgets was correctly propagated. So that more complex
widgets upon being disabled actually also disable their children. This
leads to more consistent behavior.

However, for this disabled state `Wt-disabled` was used. This is a style
class specifically for the Wt themes, not the Bootstrap themes. Now the
correct theme is queried for the disabled class name to use (in case of
Bootstrap this is `disabled`).