WT-13104: Correct Bootstrap 5 modal collision
In Bootstrap 5 the class for the modal was changed a little. This lead
to the wrapping container inside the container that Wt creates to be
caught in the maximum z-index detection as well.

Wt manages the `Wt-popup modal Wt-dialog` class, and for Bootstrap the
`modal-dialog` is added as its direct child.

While this was actually already necessary for other Bootstrap versions
too, but not correctly implemented. In Bootstrap 5 the correct chain of
DOM elements was created, as per Bootstrap's documentation. But this
lead to Wt detecting the first child of the actual dialog as a dialog
too. This had no z-index, resulting in a `NaN`.

A `Nan` value, as per JavaScript's documentation in a `Math.x` call,
will result in `NaN` being returned (see:
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Math/max).