WT-13115: Skip setDisabled optimization if parent widget changed
When a parent widget is disabled, it will propagate this to its
children. The child widgets themselves won't get a disabled flag, which
allows them to be restored to their original state when the parent
widget is re-enabled.

There is an issue when the parent widget changes. The child widget will
still have a disabled style class. When we try to enable the widget
again, the optimization of the 'setDisabled' method will early return
since the child widget was already disabled.

We can set a flag when the parent widget changed. The 'setDisabled'
method must determine the correct enabled/disabled state when the parent
widget changed.