WT-13082: make layout adjust on show
Layout does not take into account hidden elements when adjusting itself.
This is a good thing as it avoid many useless computation. However, this
also means that it needs to readjust itself when an element is that was
previously hidden is now shown. This was not done and lead to a bug.

The layout will now adjust itself correctly when a new item is shown.