WT-13063: Fix segfault for element with OTHER tags
When a DomElement was created through js, the case where the DomElement
type was OTHER was handled like the any other DomElement type, which
leads to a segfault.

This is now correctly handle, but only some tags are allowed. This is to
prevent security issues if an attacker somehow has a way to set the type
of a dom element.