WT-13163: improve touch selection in WTableView
This improve the way we handle selection in the WTableView for
touchscreen users. The problem was that it was relying on timing in
order to test if a user was scrolling or panning. This cannot be
reliable because the condition for touchmove event to be emited are not
defined.

For this reason, selection of a single element will only happen when we
recieve a touchend event and no touchstart or touchmove event were
emitted since the last touchstart signal with one touch. Also, the
selection of multiple element will now be cancel if a touchmove event is
recieve.