WT-13188: allow MIME type customization
Allows to customize the file extension to MIME type mapping used for
static resources. This can be done by specifying the path to the
wt_extMap.csv file with the --mime-map-append or
--mime-map-override in the run command.

The wt_extMap.csv file must be a CSV file containing two columns. The
first column must contain the file extension and the second column must
contain the MIME type.