WT-13300: Ensure that window state management has correct fallback
In case of a rewritten internal path (to obscure something sensitive),
no correct state can be found in Wt's internal state map.

Due to a previous check of `newState == null` this would lead to the
fallback of relying on the browser to kick in, in both the cases where
`newState` is `null` or `undefined`.

In WT-10967 this check was changed, due to the addition of a linter to a
strict type check (`newState === null`). This no longer succeeds when it
is `undefined`. This fallback shouldn't kick in often, but should still
be available in case of rewritten history.