WT-13125: move code to StdLayoutImpl::update
Move duplicate code in StdGridLayoutImpl2::update and
FlexLayoutImpl::update to StdLayoutImpl::update. This is usefull by
itself since it avoid code duplication, but also, this code must be
written in a class friend with WContainerWIdget. Since this call is
usefull as it will trigger a call to StdLayoutImpl::updateDom making it
available to any subclass of StdLayoutImpl seems to be a good thing.