WT-12967: change favicon update condition
WFavicon was notifying WApplication that the url was updated trough the
urlUpdated signal. Wapplication would the JavaScript needed to update
the favicon every time the signal was emited. This could lead to several
update of the favicon in one update, making the browser fetch favicon
that did not need to be fetch.

This solves this issue by making WApplication check if the url was
changed before sending an update to the browser instead of listnening to
the urlUpdated signal from WFavicon. This also allows to simplify
WFavicon since the urlUpdated signal is not needed anymore and is
removed.