WT-13366: Add a fallback in `WRasterImage` for docroot images
It used to be that examples from Wt were deployed such that the working
directory and the docroot of an application were the same directory.
In this case, no issue was visible. However, if the two directories
differ, we get an issue with server-side images.

The idea here is that if the client does not support JS, and cannot
dynamically update content on the page, we perform this server-side and
send the result to the client as a blob.

For images specifically this means that a `<canvas>` can be created, and
images, line, and shapes can be added to it dynamically. This is
facilitated with JavaScript (see `WPaintedWidget.js`). When the images
is added to the widget (with `drawImage()`), the path to the image is
given. This will most often be an image found in the docroot, and its
path will be specified relative to it. The browser is then able to see
the relative image URL, and fetch it from the server from its static
files.

When the fallback would occur, however, this path is interpreted as
RELATIVE to the WORKING DIRECTORY of the application, not its docroot.
Meaning the context of the two relative paths differs.

The server will NOT be able to find the image, in case the docroot is
NOT the same as the working directory. Making the fallback fail.

A more generic solution should be implemented (see WT-13367). For now
this "resolves" the issue in the widget gallery (and potentially more
places), by ensuring the relative path, if it cannot be resolved, is
also checked against the docroot.