WT-10954: add break in limitQuery in RowNum case
It was not tested as it was we do not have an Oracle database backend,
but logically, this fallthrough does not make sense since the sql added
in this case should effectively limit the size of the query. While
adding the sql in the next case may not always break the request, it is
useless at most since both limit the query.

Originally, this case had no break either, but it had no impact as the
case after it was the NotSupported case which simply break. So the
falltrough was not a problem.

This changed when the OffsetFetch case was added for SQL server. The
RowNum case was not updated at all but no break added at it's end. This
looks more like an oversight than a correction or an improvement for
Oracle database.