WT-13387: Fix MFA auth token wrong expiration date
The MFA auth token had a wrong expiration date in the database due to
Auth::AuthService always assuming the token created with
Auth::AuthService::createToken() was a basic auth token.

This is solved by adding a parameter to Auth::AuthService::createToken()
that gives the duration before the token expires.

Another option was to add a parameter telling if the auth token was an
MFA token or not, but this would lock developers into choosing between
MFA or normal duration.