JWT-13409: Allow build artifacts to be downloaded
The artifacts in question are the two resulting jar files (jwt and
jwt-auth).

This was added because JWt doesn't have any "intermediary" builds
between releases. This can be added later on, but requires changes to
the release process.

For now this is hooked into the pipeline, where JWt jars would be build
regardless. They can now be downloaded.

A manual configuration on the builder is to be performed as well, so
that a nightly build of master is performed.