WT-13410: Switch Boost archive URL
JFrog is no longer offering webspace to Boost. They used to do this, but
this is costing them a lot. Boost now offers their own in-house URL to
download archives.

See:
- https://github.com/boostorg/boost/issues/843#issuecomment-2567034014
- https://github.com/boostorg/website/pull/900

P.S. while this isn't breaking for the builds, as this is cached. This
can result in breaking builds later, if the Dockerfile were to change.