WT-13367: Add WAbstractDataInfo
When WPainter::drawImage() is called, the information needed to identify
the image are given via a string. The content expected in the string
varies depending on what paint device is used to draw the image. Some
expect a uri, other a file path to the image or a data uri. This is
problematic when using a WPaintedWidget since you do not know in advance
what paint device will be used. While this is kind of handle by assuming
the file is in the docroot, this does not leave much control to the
developper.

This solve the issue by introducing WAbstractDataInfo. Which allows to
define both a uri and a file path. Using an abstract class allows to
compute this mapping only when needed if the developper wich to do so,
by defining WAbstractDataInfo::uri() and WAbstractDataInfo::filePath().

We do define every function by default because it is possible that for
some subclasses, having a uri or a file path does not make sense, or is
irrelevent.